Move schedule collection view into its own file
